Create SolarTracker
The aim of this project is to consume the maximum solar energy through solar panel. A Solar Tracker is a device onto which solar panels are built-in which tracks the motion of the sun ensuring that maximum amount of sunlight strikes the panels all over the day. Power output from a solar cell will be maximum when it is facing the sun i.e. the angle between its surface and sun rays is 90 degrees. Solar tracking allows more energy to be produced because the solar array is able to remain aligned to the sun. The components used for its construction are servo motor, Arduino and LDR. The active sensors continuously monitor the sunlight and alternate the panel towards the direction where the intensity of sunlight is maximum. In this project, it’s divided by two categories; hardware and software. In hardware part, 2 light dependent resistor (LDR) has been used to trace the synchronize of sunlight by detecting brightness level of sunlight.